Rename `misc.hpp` to `utils.hpp`.
