+++ source/tit/core/multivector.hpp
@@ -16,7 +16,8 @@
 #include "tit/core/basic_types.hpp"
 #include "tit/core/checks.hpp"
 #include "tit/core/mdvector.hpp"
-#include "tit/core/misc.hpp"
+#include "tit/core/utils.hpp"
+
 #include "tit/par/atomic.hpp"
 #include "tit/par/thread.hpp"
