+++ source/tit/core/utils.hpp
