+++ source/tit/core/vec_simd.hpp
@@ -12,8 +12,8 @@
 #include "tit/core/basic_types.hpp"
 #include "tit/core/checks.hpp"
 #include "tit/core/math.hpp"
-#include "tit/core/misc.hpp" // IWYU pragma: keep
 #include "tit/core/simd.hpp"
+#include "tit/core/utils.hpp" // IWYU pragma: keep
 #include "tit/core/vec.hpp"
 
 namespace tit {
